requirements.txt
asgiref==3.5.2
autopep8==1.6.0
Babel==2.10.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.5
django-language-field==0.0.3
django-money==2.1.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-gis==1.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
json-tricks==3.15.5
jsonpickle==2.2.0
jsons==1.6.3
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
psycopg2-binary==2.9.3
py-moneyed==1.2
pycodestyle==2.8.0
pyparsing==3.0.9
pytz==2022.1
PyYAML==6.0
requests==2.28.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
sqlparse==0.4.2
toml==0.10.2
typish==1.9.3
uritemplate==4.1.1
urllib3==1.26.9
